collect_data/requirements.txt
aiohttp==3.5.4
google-cloud-storage==1.25.0
airly
pandas
pathlib
asyncio
